Clarify BackfillCoordinator split and wire shared service seams
- Move Meridian.Application BackfillCoordinator from Http/ (namespace
  Meridian.Application.UI) to Backfill/ (namespace
  Meridian.Application.Backfill) so folder, namespace, and role agree.
- Move preview/estimation logic from the Ui.Shared wrapper down into the
  core coordinator (new BackfillPreviewPlanner + PreviewAsync); the
  Ui.Shared BackfillCoordinator is now a thin read-model facade and
  wraps the composition-root core singleton when registered so both
  surfaces share one run gate.
- Implement ICredentialService and IWatchlistService on the WPF
  CredentialService and WatchlistService via explicit adapter members,
  and align the shared no-op defaults' doc comments with the real seam
  (Notification/Schema families already ride NotificationServiceBase/
  SchemaServiceBase).
