+++ src/Meridian.Application/Composition/Features/StorageFeatureRegistration.cs
@@ -1,3 +1,4 @@
+using Meridian.Application.Backfill;
 using Meridian.Core.Config;
 using Meridian.PortfolioRecords.Accounts;
 using Meridian.FinancialOperations.Banking;
