+++ src/Meridian.Wpf/Services/CredentialService.cs
@@ -12,8 +12,14 @@ using System.Threading;
 using System.Threading.Tasks;
 using Meridian.Contracts.Credentials;
 using Meridian.Ui.Services;
+using Meridian.Ui.Services.Contracts;
 using HttpClientFactoryProvider = Meridian.Ui.Services.HttpClientFactoryProvider;
 using HttpClientNames = Meridian.Ui.Services.HttpClientNames;
+using SharedCredentialExpirationEventArgs = Meridian.Ui.Services.CredentialExpirationEventArgs;
+using SharedCredentialMetadataInfo = Meridian.Ui.Services.CredentialMetadataInfo;
+using SharedCredentialMetadataUpdate = Meridian.Ui.Services.CredentialMetadataUpdate;
+using SharedCredentialWithMetadata = Meridian.Ui.Services.CredentialWithMetadata;
+using SharedOAuthRefreshResult = Meridian.Ui.Services.OAuthRefreshResult;
 
 namespace Meridian.Wpf.Services;
 
@@ -106,8 +112,10 @@ public sealed class CredentialExpirationEventArgs : EventArgs
 /// <summary>
 /// Service for secure credential management using DPAPI-encrypted file storage.
 /// Enhanced with OAuth support, expiration tracking, and credential testing capabilities.
+/// Implements the shared <see cref="ICredentialService"/> seam so shared UI services can
+/// consume this host implementation instead of the no-op default.
 /// </summary>
-public sealed class CredentialService : IDisposable
+public sealed class CredentialService : ICredentialService, IDisposable
 {
     private const string ResourcePrefix = "Meridian";
     private const string CredentialVaultFileName = "credentials.enc";
@@ -502,6 +510,7 @@ public sealed class CredentialService : IDisposable
             if (remaining.TotalDays <= 7 && remaining.TotalSeconds > 0)
             {
                 CredentialExpiring?.Invoke(this, new CredentialExpirationEventArgs(resource, metadata.ExpiresAt.Value));
+                _sharedCredentialExpiring?.Invoke(this, new SharedCredentialExpirationEventArgs(resource, metadata.ExpiresAt.Value));
             }
         }
     }
@@ -928,6 +937,66 @@ public sealed class CredentialService : IDisposable
     }
 
 
+
+    // ICredentialService adapter: explicit implementations map the narrower shared UI seam
+    // onto this host implementation, so shared services get real credentials without the
+    // WPF surface (richer metadata, typed test results) leaking into the contract.
+
+    private EventHandler<SharedCredentialExpirationEventArgs>? _sharedCredentialExpiring;
+
+    event EventHandler<SharedCredentialExpirationEventArgs>? ICredentialService.CredentialExpiring
+    {
+        add => _sharedCredentialExpiring += value;
+        remove => _sharedCredentialExpiring -= value;
+    }
+
+    IReadOnlyList<SharedCredentialWithMetadata> ICredentialService.GetAllCredentialsWithMetadata()
+        => GetAllCredentialsWithMetadata()
+            .Select(c => new SharedCredentialWithMetadata
+            {
+                Resource = c.Resource,
+                IsOAuthToken = c.CredentialType == CredentialType.OAuth2Token,
+                ExpiresAt = c.ExpiresAt,
+                CanAutoRefresh = c.CanAutoRefresh,
+                AutoRefreshEnabled = c.CanAutoRefresh
+            })
+            .ToList();
+
+    async Task<SharedOAuthRefreshResult> ICredentialService.RefreshOAuthTokenAsync(string providerId, CancellationToken ct)
+    {
+        var success = await RefreshOAuthTokenAsync(providerId, ct).ConfigureAwait(false);
+        var metadata = GetMetadata($"{OAuthTokenResource}.{providerId}");
+        return new SharedOAuthRefreshResult
+        {
+            Success = success,
+            ErrorMessage = success ? null : "OAuth token refresh failed.",
+            NewExpiration = success ? metadata?.ExpiresAt : null
+        };
+    }
+
+    Task ICredentialService.UpdateMetadataAsync(string resource, Action<SharedCredentialMetadataUpdate> updateAction, CancellationToken ct)
+    {
+        ArgumentNullException.ThrowIfNull(updateAction);
+
+        var update = new SharedCredentialMetadataUpdate
+        {
+            AutoRefreshEnabled = GetMetadata(resource)?.AutoRefreshEnabled ?? false
+        };
+        updateAction(update);
+        return UpdateMetadataAsync(resource, m => m.AutoRefreshEnabled = update.AutoRefreshEnabled, ct);
+    }
+
+    SharedCredentialMetadataInfo? ICredentialService.GetMetadata(string resource)
+    {
+        var metadata = GetMetadata(resource);
+        return metadata is null
+            ? null
+            : new SharedCredentialMetadataInfo
+            {
+                AutoRefreshEnabled = metadata.AutoRefreshEnabled,
+                ExpiresAt = metadata.ExpiresAt
+            };
+    }
 
     private void LogCredentialOperation(string operation, string resource, bool success, string? details = null)
     {
