+++ tests/Meridian.Tests/Application/Backfill/BackfillCoordinatorStorageOptionsTests.cs
@@ -1,10 +1,9 @@
 using FluentAssertions;
-using Meridian.Application.UI;
+using Meridian.Application.Backfill;
 using Meridian.Core.Config;
 using Meridian.Infrastructure.Adapters.Core;
 using Meridian.Storage;
 using Xunit;
-using BackfillRequest = Meridian.Application.Backfill.BackfillRequest;
 
 namespace Meridian.Tests.Application.Backfill;
